Fix int32 offset overflow in NLLLoss2d reduce backward kernel
nll_loss2d_backward_kernel computed the per-sample base offsets sample,
toffset, and ioffset as 32-bit int. ioffset = sample * map_nelem * n_classes
therefore wrapped once grad_input had more than INT_MAX elements, so
grad_input_thread pointed outside the allocation and the gradient store issued
an illegal memory access. This path is reached by cross_entropy on
(batch, classes, d1, ...) logits, i.e. ordinary large-batch spatial losses,
not only a synthetic case. The forward and no-reduce kernels already carry
these offsets in the templated index_t; widen the reduce kernel's three
offset locals to int64_t so the products are computed in 64-bit.

Test Plan:
Added a large-tensor backward regression (grad_input numel > INT_MAX) that
routes cross_entropy through the 2d reduce backward kernel:

```
python test/test_nn.py -k test_nll_loss2d_backward_64bit_offsets
```

+++ aten/src/ATen/native/cuda/NLLLoss2d.cu
@@ -182,13 +182,13 @@ __global__ void nll_loss2d_backward_kernel(
   const auto grad = -(size_average ? *grad_output / *total_weight
                                    : *grad_output);
 
-  const int sample = blockIdx.x / blocks_per_sample;
+  const int64_t sample = blockIdx.x / blocks_per_sample;
   const int step = blockDim.x * blocks_per_sample;
 
-  const int toffset = sample * map_nelem;
+  const int64_t toffset = sample * map_nelem;
   const auto* const target_thread = target + toffset;
 
-  const int ioffset = sample * map_nelem * n_classes;
+  const int64_t ioffset = sample * map_nelem * n_classes;
   auto* const grad_input_thread = grad_input + ioffset;
 
   for (int i = (blockIdx.x % blocks_per_sample) * blockDim.x + threadIdx.x;
